+++ providers/gcs/gcs.go
@@ -194,7 +194,7 @@ func (b *Bucket) Name() string {
 }
 
 func (b *Bucket) SupportedIterOptions() []objstore.IterOptionType {
-	return []objstore.IterOptionType{objstore.Recursive, objstore.UpdatedAt}
+	return []objstore.IterOptionType{objstore.Recursive, objstore.UpdatedAt, objstore.StartAfter}
 }
 
 func (b *Bucket) IterWithAttributes(ctx context.Context, dir string, f func(attrs objstore.IterObjectAttributes) error, options ...objstore.IterOption) error {
@@ -217,8 +217,9 @@ func (b *Bucket) IterWithAttributes(ctx context.Context, dir string, f func(attr
 	}
 
 	query := &storage.Query{
-		Prefix:    dir,
-		Delimiter: delimiter,
+		Prefix:      dir,
+		Delimiter:   delimiter,
+		StartOffset: appliedOpts.StartAfter,
 	}
 	if appliedOpts.LastModified {
 		if err := query.SetAttrSelection([]string{"Name", "Updated"}); err != nil {
@@ -257,12 +258,12 @@ func (b *Bucket) IterWithAttributes(ctx context.Context, dir string, f func(attr
 // Iter calls f for each entry in the given directory. The argument to f is the full
 // object name including the prefix of the inspected directory.
 func (b *Bucket) Iter(ctx context.Context, dir string, f func(string) error, opts ...objstore.IterOption) error {
-	// Only include recursive option since attributes are not used in this method.
+	// Only include supported options since attributes are not used in this method.
 	var filteredOpts []objstore.IterOption
 	for _, opt := range opts {
-		if opt.Type == objstore.Recursive {
+		switch opt.Type {
+		case objstore.Recursive, objstore.StartAfter:
 			filteredOpts = append(filteredOpts, opt)
-			break
 		}
 	}